refactor(web): 프론트의 Twilio 네이밍을 휴대폰 인증으로 정리
발송 벤더가 NHN Cloud SMS 로 바뀌었는데 프론트 심볼이 옛 벤더 이름을 달고 있었다.

storefront:
  lib/api/users/twilio/        → lib/api/users/phone-verification/
  hooks/use-twilio.ts          → hooks/use-phone-verification.ts
  sendTwilioMessageApi         → sendVerificationCodeApi
  SendTwilioMessageDto         → SendVerificationCodeDto
  useTwilio / sendTwilioMessage / resetTwilio
                               → usePhoneVerification / sendVerificationCode /
                                 resetPhoneVerification

auth-web 은 함수명이 이미 sendPhoneVerificationCode·verifyPhoneCode 라 그대로 두고,
"Twilio Verify 는 approve 된 코드를 재검증할 수 없다"는 낡은 근거만 지웠다 —
verify-code.service 의 alreadyVerified 가 재검증을 허용하므로 더는 사실이 아니다.
가입 시점에만 검증한다는 보안 논지는 남겼다.

서버 쪽 errorCode 의 TWILIO_ 접두도 뗐다(NO_VALID_VERIFICATION,
MAX_ATTEMPTS_EXCEEDED). 직전 커밋 전까지 getErrorCode() 가 늘 undefined 를 반환해
응답에 실린 적이 없으므로 이 이름에 의존하는 클라이언트는 없다.

개인정보처리방침도 맞춘다. 제6조 국외 이전에서 Twilio Inc. 항목을 빼고, 제5조 위탁의
엔에이치엔클라우드 업무 범위를 "카카오 알림톡"에서 "카카오 알림톡 및 인증·안내 문자(SMS)"
로 넓혔다. SMS 가 국내 발신망으로 바뀌면 미국으로의 국외 이전이 사라지기 때문이다.
consents/ 와 checkout/policies/ 두 사본 모두 적용. 시행일은 건드리지 않았다 — 개정 공지
여부는 별도 판단이 필요하다.

라우트 문자열 /twilio/send-message·/twilio/verify-code 는 유지한다. 서버와 프론트 둘을
같은 순간에 배포해야 하는 변경이라 경로 정리는 따로 한다.

⚠️ 처리방침 변경이 들어있으므로 실제 NHN 전환 배포와 함께 나가야 한다. storefront 만
먼저 배포하면 Twilio 를 쓰면서 안 쓴다고 고지하는 상태가 된다.
